+++ src/core/index.ts
@@ -45,6 +45,9 @@ export type {
 export { ExpandableBlock } from './ExpandableBlock';
 export type { ExpandableBlockProps } from './ExpandableBlock';
 
+export { FileUpload, FileUploadTemplate } from './FileUpload';
+export type { FileUploadProps, FileUploadTemplateProps } from './FileUpload';
+
 export { Footer } from './Footer';
 export type { FooterProps, FooterElement, TitleTranslations } from './Footer';
 
@@ -87,8 +90,13 @@ export type { RadioTileGroupProps, RadioTileProps } from './RadioTiles';
 export { Select } from './Select';
 export type { SelectProps, SelectOption, ItemRendererProps } from './Select';
 
-export { Table } from './Table';
-export type { TableProps } from './Table';
+export { Table, tableFilters, FilterButtonBar } from './Table';
+export type {
+  TableProps,
+  TableFilterProps,
+  TableFilterValue,
+  FilterButtonBarProps,
+} from './Table';
 
 export { Tag, TagContainer } from './Tag';
 export type { TagProps, TagContainerProps } from './Tag';